+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### Unreleased
 
+### [v0.10](https://github.com/arez/arez-idlestatus/tree/v0.10) (2018-05-02)
+[Full Changelog](https://github.com/arez/arez-idlestatus/compare/v0.09...v0.10)
+
+* Update the 'org.realityforge.arez' dependencies to version '0.82'
+
 ### [v0.09](https://github.com/arez/arez-idlestatus/tree/v0.09) (2018-04-27)
 [Full Changelog](https://github.com/arez/arez-idlestatus/compare/v0.08...v0.09)